debug d

+++ submodules/diff-surfel-sdf-rasterization/cuda_rasterizer/forward.h
@@ -67,7 +67,7 @@ namespace FORWARD
 		const float* bg_color,
 		float* out_color,
 		float* out_others,
-		float* gs_per_pixel,
+		int* gs_per_pixel,
 		float* weight_per_gs_pixel,
 		float* x_mu
 	);
