+++ submodules/diff-surfel-sdf-rasterization/cuda_rasterizer/rasterizer.h
@@ -51,7 +51,7 @@ namespace CudaRasterizer
 			const bool prefiltered,
 			float* out_color,
 			float* out_others,
-			float* gs_per_pixel,
+			int* gs_per_pixel,
 			float* weight_per_gs_pixel,
 			float* x_mu,
 			int* radii = nullptr,
